added hardcoded GOT problem #1

+++ src/searching/problems/SearchAction.java
@@ -1,14 +1,14 @@
 package searching.problems;
 
 public interface SearchAction {
-	public int getCost();
+	public long getCost();
 	public static final class NoAction implements SearchAction {
 		public NoAction() {
 			super();
 		}
 
 		@Override
-		public int getCost() {
+		public long getCost() {
 			return 0;
 		}
 	} //used for root nodes
